# BACK/swagger.yaml
openapi: 3.0.0
info:
  title: API Doceria Gourmet IANES
  description: >
    API RESTful para gerenciamento de estoque da Doceria Gourmet IANES.
    Controla produtos, lotes, movimentações e notificações de validade e estoque baixo.
  version: 1.0.0
  contact:
    name: Sidney Campos Aragão
servers:
  - url: http://localhost:8080/v1/IANES
    description: Ambiente de desenvolvimento
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErroValidacao:
      type: object
      properties:
        status:       { type: string,  example: "Bad Request" }
        status_code:  { type: integer, example: 400 }
        message:      { type: string,  example: "Campos obrigatórios não preenchidos corretamente." }
    ErroNaoEncontrado:
      type: object
      properties:
        status:       { type: string,  example: "Not Found" }
        status_code:  { type: integer, example: 404 }
        message:      { type: string,  example: "Nenhum item encontrado." }
...
paths:
...
  /produto:
    get:
      tags: [Produto]
      summary: Listar todos os produtos
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Produtos retornados com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  produtos:
                    - id_produto: 1
                      nome_produto: "Bolo de Chocolate"
                      codigo_identificador: 1001
                      sabor_massa: "Chocolate"
                      recheio: "Brigadeiro"
                      cobertura: "Ganache"
                      detalhes: "Sem glúten"
                      limite_minimo_alerta: 3
                      id_categoria: 1
                      nome_categoria: "Bolo"

    post:
      tags: [Produto]
      summary: Cadastrar novo produto
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            example:
              nome_produto: "Bolo de Chocolate"
              id_categoria: 1
              sabor_massa: "Chocolate"
              recheio: "Brigadeiro"
              cobertura: "Ganache"
              detalhes: "Sem glúten"
              limite_minimo_alerta: 3
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              example:
                status: "Created"
                status_code: 201
                message: "Item criado com sucesso!"
                response:
                  produto:
                    id_produto: 1
                    nome_produto: "Bolo de Chocolate"
                    codigo_identificador: 1001
                    id_categoria: 1
        '400': { description: Dados inválidos, content: { application/json: { schema: { $ref: '#/components/schemas/ErroValidacao' } } } }
        '404': { description: Categoria não encontrada, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoEncontrado' } } } }
  /produto/filtro:
    get:
      tags: [Produto]
      summary: Buscar produtos com filtros dinâmicos
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: nome
          schema: { type: string, example: "bolo" }
        - in: query
          name: codigo
          schema: { type: integer, example: 1001 }
        - in: query
          name: categoria
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Produtos encontrados
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  produtos:
                    - id_produto: 1
                      nome_produto: "Bolo de Chocolate"
                      codigo_identificador: 1001
                      nome_categoria: "Bolo"
        '400': { description: Nenhum filtro informado, content: { application/json: { schema: { $ref: '#/components/schemas/ErroValidacao' } } } }
  /produto/proximo-codigo:
    get:
      tags: [Produto]
      summary: Retorna o próximo código identificador disponível
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Próximo código retornado
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  proximo_codigo: 1002
  /produto/{id}:
    get:
      tags: [Produto]
      summary: Buscar produto por ID
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  produto:
                    id_produto: 1
                    nome_produto: "Bolo de Chocolate"
                    codigo_identificador: 1001
                    nome_categoria: "Bolo"
        '404': { description: Produto não encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoEncontrado' } } } }

    put:
      tags: [Produto]
      summary: Atualizar produto (parcial)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      requestBody:
        required: true
        content:
          application/json:
            example:
              recheio: "Doce de Leite"
              limite_minimo_alerta: 5
      responses:
        '200':
          description: Produto atualizado com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Item atualizado com sucesso!"
                response:
                  produto:
                    id_produto: 1
                    recheio: "Doce de Leite"
                    limite_minimo_alerta: 5

    delete:
      tags: [Produto]
      summary: Excluir produto (cascata nos lotes)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Produto excluído com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Item deletado com sucesso!"
        '404': { description: Produto não encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoEncontrado' } } } }

  # ══════════════════════════════════════════════
  # LOTE
  # ══════════════════════════════════════════════
...
  /lote/{id}:
    get:
      tags: [Lote]
      summary: Buscar lote por ID
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Lote encontrado
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  lote:
                    id_lote: 1
                    status_validade: "No prazo"
                    nome_produto: "Bolo de Chocolate"

    put:
      tags: [Lote]
      summary: Atualizar lote (recalcula status automaticamente)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      requestBody:
        required: true
        content:
          application/json:
            example:
              data_vencimento: "2026-06-10"
      responses:
        '200':
          description: Lote atualizado com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Item atualizado com sucesso!"
                response:
                  lote:
                    id_lote: 1
                    data_vencimento: "2026-06-10"
                    status_validade: "No prazo"

    delete:
      tags: [Lote]
      summary: Excluir lote (falha se houver movimentações vinculadas)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Lote excluído com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Item deletado com sucesso!"
        '500':
          description: Lote possui movimentações vinculadas
          content:
            application/json:
              example:
                status: "Internal Server Error"
                status_code: 500
                message: "Não é possível excluir um lote com movimentações registradas!"

  # ══════════════════════════════════════════════
  # MOVIMENTAÇÃO
  # ══════════════════════════════════════════════
  /movimentacao:
    get:
      tags: [Movimentação]
      summary: Listar histórico completo de movimentações
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Movimentações retornadas com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  movimentacoes:
                    - id_movimentacao: 1
                      tipo_movimentacao: "Entrada"
                      motivo: "Nova fornada produzida"
                      quantidade: 12
                      data_hora: "2026-05-27T10:30:00"
                      nome_produto: "Bolo de Chocolate"
                      nome_usuario: "Sidney Aragão"

    post:
      tags: [Movimentação]
      summary: Registrar movimentação de entrada ou saída
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            example:
              tipo_movimentacao: "Saída"
              motivo: "Venda balcão"
              quantidade: 3
              id_lote: 1
      responses:
        '201':
          description: Movimentação registrada com sucesso
          content:
            application/json:
              example:
                status: "Created"
                status_code: 201
                message: "Item criado com sucesso!"
                response:
                  movimentacao:
                    id_movimentacao: 2
                    tipo_movimentacao: "Saída"
                    motivo: "Venda balcão"
                    quantidade: 3
                    data_hora: "2026-05-27T14:00:00"
                    nome_produto: "Bolo de Chocolate"
                    nome_usuario: "Sidney Aragão"
        '400':
          description: Saldo insuficiente ou lote vencido
          content:
            application/json:
              example:
                status: "Bad Request"
                status_code: 400
                message: "Quantidade de saída maior que o estoque disponível no lote."
  /movimentacao/{id}:
    get:
      tags: [Movimentação]
      summary: Buscar movimentação por ID
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Movimentação encontrada
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  movimentacao:
                    id_movimentacao: 1
                    tipo_movimentacao: "Entrada"
                    quantidade: 12
                    nome_produto: "Bolo de Chocolate"
  /movimentacao/lote/{id_lote}:
    get:
      tags: [Movimentação]
      summary: Listar movimentações de um lote específico
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id_lote
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Histórico do lote retornado
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  movimentacoes:
                    - id_movimentacao: 1
                      tipo_movimentacao: "Entrada"
                      motivo: "Nova fornada"
                      quantidade: 12
                      data_hora: "2026-05-27T10:30:00"
                      nome_usuario: "Sidney Aragão"
                    - id_movimentacao: 2
                      tipo_movimentacao: "Saída"
                      motivo: "Venda balcão"
                      quantidade: 3
                      data_hora: "2026-05-27T14:00:00"
                      nome_usuario: "Sidney Aragão"

  # ══════════════════════════════════════════════
  # NOTIFICAÇÃO
  # ══════════════════════════════════════════════
...